While studying the games of my agents, I realized that the feature planes being fed into the neural networks were holding them back. They made the networks blind to the differences between checker counts >=4. And these distinctions have strategic significance. For example, if there are at least 6 checkers at a point, then if a double is rolled, all 4 rolls can applied to checkers at that one point without making it vulnerable.
For this reason, I have changed the recommended feature planes to
include separate planes for checker counts up to 6, and an additional
plane for the normalized excess count above 6 (out of 4).

This increases the number of feature planes from 9 to 15.

This is a significant change for the network to deal with (although it
is really only adding additional weights to the first ConvNet, the
network will likely learn different internal representations of the game
state in the internal 128 channels), but not for the programmer. All you
need to do is add a few lines to your encoding function. And change the
constants CONV_INPUT_CHANNELS in your network files, which I've done in
this commit.

This could break your code, but all you need to do is add a few lines to
your encoder, or else just change the constant CONV_INPUT_CHANNELS back
to 9.